feat: update bls12381G2 names and add keylength check to the constructor
This renames all of the G2 related files and tests to be named specifically for the G2 subgroup in
preperation for supporting G1 key pairs as well. Additionally a keylength check was added to the
constructor to fail fast if the wrong length string is provided as a public key or a private key.